# docs/generated/openapi/agent.openapi.json
{
  "components": {
    "schemas": {
...
      "FunctionBulkSync": {
        "properties": {
          "functions": {
            "items": {
              "$ref": "#/components/schemas/FunctionRegister"
            },
            "maxItems": 512,
            "title": "Functions",
            "type": "array"
          }
        },
        "required": [
          "functions"
        ],
        "title": "FunctionBulkSync",
        "type": "object"
      },
      "FunctionOut": {
        "properties": {
          "app_id": {
            "format": "uuid",
            "title": "App Id",
            "type": "string"
          },
          "availability": {
            "additionalProperties": true,
            "title": "Availability",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "description_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Override"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pack_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Name"
          },
          "parameters_schema": {
            "additionalProperties": true,
            "title": "Parameters Schema",
            "type": "object"
          },
          "required_capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Required Capabilities",
            "type": "array"
          },
          "required_entitlements": {
            "items": {
              "type": "string"
            },
            "title": "Required Entitlements",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "timeout_seconds": {
            "title": "Timeout Seconds",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "app_id",
          "name",
          "description",
          "parameters_schema",
          "is_active",
          "timeout_seconds",
          "availability",
          "required_entitlements",
          "required_capabilities",
          "source",
          "created_at"
        ],
        "title": "FunctionOut",
        "type": "object"
      },
      "FunctionRegister": {
        "properties": {
          "availability": {
            "additionalProperties": true,
            "title": "Availability",
            "type": "object"
          },
          "description": {
            "default": "",
            "maxLength": 1000,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "pack_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Name"
          },
          "parameters_schema": {
            "additionalProperties": true,
            "title": "Parameters Schema",
            "type": "object"
          },
          "required_capabilities": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "title": "Required Capabilities",
            "type": "array"
          },
          "required_entitlements": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "title": "Required Entitlements",
            "type": "array"
          },
          "source": {
            "default": "app_inline",
            "pattern": "^(app_inline|playbook_pack)$",
            "title": "Source",
            "type": "string"
          },
          "timeout_seconds": {
            "default": 30,
            "maximum": 300.0,
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "type": "integer"
          }
        },
        "required": [
          "name"
        ],
        "title": "FunctionRegister",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Runtime service for Playbook SDK chat, sessions, and function execution",
    "title": "Playbook Agent Runtime API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/functions/bulk": {
      "put": {
        "operationId": "bulk_sync_functions_v1_functions_bulk_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FunctionBulkSync"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FunctionOut"
                  },
                  "title": "Response Bulk Sync Functions V1 Functions Bulk Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Sync Functions",
        "tags": [
          "functions-sdk"
        ]
      }
    },
    "/v1/functions/eligible": {
      "get": {
        "operationId": "list_eligible_functions_sdk_v1_functions_eligible_get",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FunctionOut"
                  },
                  "title": "Response List Eligible Functions Sdk V1 Functions Eligible Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Eligible Functions Sdk",
        "tags": [
          "functions-sdk"
        ]
      }
    },
...
  }
}